Make the ids mostly persistent
The ids are supposed to identify the resource between re-compiles.
So just generating a new id for the same resource each build is dumb.
Right now I'm basing the ID off of the link. I don't plan to change links often
(or at all, actually, to avoid dead links, not like anyone links to my site)
so this solution should be pretty solid, I think.

Instead of just dropping the link into the ID itself, I am hashing it, so it
looks like a uuid. This is because it is set as the `guid` property in the rss
feed, and it's pretty standard to use uuids as the id. I don't know if any feed
readers rely on it being a uuid-looking-string, (I would hope not), but just in
case, I guess.